widgets/forms directory refactored - all field classes and methods moved to separate file
small naming/import changes

WIDGETS FRNT-2 [IN PROGRESS]
Form FRNT-5 [IN PROGRESS]
