+++ BaCa2/templates/widget_templates/forms/base.html
@@ -34,7 +34,7 @@
                 url: '{{ form_widget.post_url }}',
                 data: $(this).serialize(),
                 success: function (data) {
-                    // TODO
+                    console.log('success');
                 }
             });
         });
