+++ BaCa2/widgets/forms/__init__.py
@@ -1,1 +1,2 @@
-from .base import BaCa2Form, FormWidget
+from widgets.forms.base import BaCa2Form, FormWidget
+from widgets.forms.fields import get_field_validation_status
